misc: tsconfig in libs should not point to post-compile tsconfig
Currently, in root there are 2 tsconfig files, the default and tsconfig.lib.json
The default has path mappings to the source code.
The tsconfig.lib.json has path mappings to the compile targers (d.ts in dist folder).

All tsconfig.json that are local in the root of each library source pointed to the lib.json
which causes VSCode to link to the d.ts mappings, instead it is now pointing to the tsconfig.json
and the local tsconfig.lib.json will point to the root tsconfig.lib.json