chore: refactor the file structure
This re-structure does not involve files related to the libraries, it has effect on all the things around them, such as internal build librarys, demo app feature modules etc. Up to now all of the libraries were in the `libs` folder which means taht a public library is in the same location as an internal build libarary for the project or feature module libraries for the demo app.