feat(ngrid): new api to add/remove columns refactor(ngrid) remove hideColumns input from ngrid and use a directive for it feat(ngrid): new api to control rows (RowsApi) feat(ngrid): new onConstructed event fired after the grid is constructed but before it is initialized
BREAKING CHANGE:
If you used the hideColumns property (setter only) via code and not via html binding it will no longer work.
Instead, use the new api to add/remove columns.
If you used it via bindings, it will still work but it is not recommended because other plugins that use the API
will override values from the array provided.