build(ngrid): refactor import/export structure in files
The refactor in previous commit broke the library builds.
This was expected because a lot of files and their import/export order has changed.
Angular package format buliders, using NGC, are very sensetive to this kind of thing.
This commit fixes this by removing most of the barrel imports in the main modules (ngrid module and component)