feat(ngrid): allow minimum height based on row count
BREAKING CHANGES:

`PblNgridComponent.fallbackMinHeight` is now deprecated and will be removed in the next major version (4.0.0), use `PblNgridComponent.minDataViewHeight` instead

The UI layout has changed when virtual scroll is disabled (`vScrollNone`).
Previously, the layout was inconsistent, creating different behviour when vScroll was disabled. The fixed metadata rows were set within the container and the data viewport had no scrollbar. This is now changed and it will behave exactly the same in both modes.