chore(apps): restructure shared app libs to new state, having 2 apps
Up to this point we had 2 application and there was a section for application libs
that the app used. It was a clutter since we only had 1 app, no issues surfaced.
Now, with 2 apps, we need a little bit of code sharing, so we need to refactor the libs
to support that.