perf(ngrid): use internal row context
With this commit, the context provided to rows is now an internal ngrid context class and not the context provided by CdkTable.
This allow better control over the context but comes with a big refactor removing large chunks of code used to sync between the previous added context and the CdkContext.