feat(ngrid/detail-row): implement global detail row instance manager
feat(ngrid): implement dynamic virtual scroll strategy

The global detail row manager is responsible for managing detail row instances.
From creationg until destruction it is all managed in the detail row controller.
The detail row parents are now only handling change in state and in-turn pass the
required operation to the detail row manager.
All of the detail rows are created under the same view container ref which means
that we can pass them between rows and save DOM elements and view refs, just change the context.

The dynamic virtual scroll strategy is a new strategy more suitable for nGrid
allowing dynamic row heights to be used and thus supporting detail rows in virtual scroll.
It is something in between the fixed and the auto strategies.
Currently, the fixed and auto strategies are also avilable but they will be deprecated from the main package in the next version and moved to a package of their own.