feat(ngrid/detail-row): implement global detail row instance manager
The global detail row manager is responsible for managing detail row instances.
From creationg until destruction it is all managed in the detail row controller.
The detail row parents are now only handling change in state and in-turn pass the
required operation to the detail row manager.
All of the detail rows are created under the same view container ref which means
that we can pass them between rows and save DOM elements and view refs, just change the context.