feat(ngrid): implement dynamic virtual scroll strategy
The dynamic virtual scroll strategy is a new strategy more suitable for nGrid
allowing dynamic row heights to be used and thus supporting detail rows in virtual scroll.
It is something in between the fixed and the auto strategies.
Currently, the fixed and auto strategies are also avilable but they will be deprecated from the main package in the next version and moved to a package of their own.