revert(ngrid): PblNgridRegistryService back to main package
It seems that `ngcc` does not fancy having an injectable in the core package. This is true even if the core package epxoses it's own ngModule. It's probably that the registry service is a provider in every grid instance and it injects itself into it's own constructor. This seems to be too much for `ngcc`.
We will have to wait with this move to the point where ngcc is no longer a thing and viewEngine is dead