chore: faster dev builds
The dev build we're super slow, approx 4 minutes for startup and 60 seconds between save & reload.
This is due to several issues which were fixed in this commit.

1) The default for `extractLicenses` is true which means it  built the licenses every save & reload
It is now completely removed in dev server buidls.

2) There is A metadata plugin that extracts the package versions from package.json files (ng, cdk, ngrid) and the latest git SHA.
This plugin had a huge impact on performance because it did a git query for every child compilation (i.e. all HTML, SCSS files).
In addition, on every save & reload a new DefinePlugin was created and added, alongside existing onces, which bloated the memory and slowed the entire build. worse plugin hack ever!
Now, it will only run once, at startup.

3) There are multiple webpack plugins which runs the entire process.
The main 2 are responsible for markdown to HTML and extracting code examples into HTML.
Things changed in the angular CLI and webpack integration since they we're built and they required refactor because they we're extremly slow.
This is now in place.

Before this commit we have around 60 seconds between save & reload, not it's approx. 4-5 seconds.
There is still room for optimizations but it should be fine for now.